+++ tests/General.py
@@ -1,12 +1,12 @@
 from teddecor.UnitTest import *
 
 
-@test
-def _12_eq_12():
-    assertThat(12, eq(12))
+class Testing(Test):
+    @test
+    def _12_eq_12(self):
+        assertThat(12, eq(12))
 
 
 if __name__ == "__main__":
-    # assertThat(None, _is(12))
-    run(_12_eq_12).save(ext=SaveType.TXT)
-    # print(TED.strip("\[[@F red]Error Message[@F]] \x1b[1m*Some error \\_message"))
+    TestSuite("Suite Filtering", tests=[Testing]).run(filter=[TestFilter.TOTALS])
+    # Testing().run(filter=[TestFilter.TOTALS, TestFilter.SKIPPED])